+++ xknx/dpt/dpt_1byte_signed.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from xknx.exceptions import ConversionError
 
 from .dpt import DPTNumeric
-from .payload import DPTArray
+from .payload import DPTArray, DPTBinary
 
 
 class DPTSignedRelativeValue(DPTNumeric):
@@ -24,9 +24,9 @@ class DPTSignedRelativeValue(DPTNumeric):
     resolution = 1
 
     @classmethod
-    def from_knx(cls, raw: tuple[int, ...]) -> int:
+    def from_knx(cls, payload: DPTArray | DPTBinary) -> int:
         """Parse/deserialize from KNX/IP raw data."""
-        cls.test_bytesarray(raw)
+        raw = cls.validate_payload(payload)
         if raw[0] > cls.value_max:
             return raw[0] - 0x100
         return raw[0]