+++ xknx/dpt/dpt_1byte_uint.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from xknx.exceptions import ConversionError
 
 from .dpt import DPTNumeric
-from .payload import DPTArray
+from .payload import DPTArray, DPTBinary
 
 
 class DPTValue1ByteUnsigned(DPTNumeric):
@@ -24,15 +24,13 @@ class DPTValue1ByteUnsigned(DPTNumeric):
     resolution = 1
 
     @classmethod
-    def from_knx(cls, raw: tuple[int, ...]) -> int:
+    def from_knx(cls, payload: DPTArray | DPTBinary) -> int:
         """Parse/deserialize from KNX/IP raw data."""
-        cls.test_bytesarray(raw)
-
-        value = raw[0]
+        value = cls.validate_payload(payload)[0]
 
         if not cls._test_boundaries(value):
             raise ConversionError(
-                f"Could not parse {cls.__name__}", value=value, raw=raw
+                f"Could not parse {cls.__name__}", value=value, payload=payload
             )
 
         return value
@@ -120,15 +118,13 @@ class DPTSceneNumber(DPTValue1ByteUnsigned):
     value_max = 64
 
     @classmethod
-    def from_knx(cls, raw: tuple[int, ...]) -> int:
+    def from_knx(cls, payload: DPTArray | DPTBinary) -> int:
         """Parse/deserialize from KNX/IP raw data."""
-        cls.test_bytesarray(raw)
-
-        value = raw[0] + 1
+        value = cls.validate_payload(payload)[0] + 1
 
         if not cls._test_boundaries(value):
             raise ConversionError(
-                f"Could not parse {cls.__name__}", value=value, raw=raw
+                f"Could not parse {cls.__name__}", value=value, payload=payload
             )
 
         return value