+++ xknx/dpt/dpt_scaling.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from xknx.exceptions import ConversionError
 
 from .dpt import DPTNumeric
-from .payload import DPTArray
+from .payload import DPTArray, DPTBinary
 
 
 class DPTScaling(DPTNumeric):
@@ -25,19 +25,16 @@ class DPTScaling(DPTNumeric):
     resolution = 1
 
     @classmethod
-    def from_knx(cls, raw: tuple[int, ...]) -> int:
+    def from_knx(cls, payload: DPTArray | DPTBinary) -> int:
         """Parse/deserialize from KNX/IP raw data."""
-        cls.test_bytesarray(raw)
-
-        knx_value = raw[0]
+        knx_value = cls.validate_payload(payload)[0]
         delta = cls.value_max - cls.value_min
         value = round((knx_value / 255) * delta) + cls.value_min
 
         if not cls._test_boundaries(value):
             raise ConversionError(
-                f"Could not parse {cls.__name__}", value=value, raw=raw
+                f"Could not parse {cls.__name__}", value=value, payload=payload
             )
-
         return value
 
     @classmethod