+++ docs/changelog.md
@@ -14,7 +14,9 @@ nav_order: 2
 
 ### Internals
 
-- Return `DPTArray` or `DPTBinary` from `DPTBase.to_knx()` instead of `tuple[int, ...]`.
+- Refactor DPTBase transcoder classes
+  - Accept `DPTArray` or `DPTBinary` in `DPTBase.from_knx()` instead of raw `tuple[int]`.
+  - Return `DPTArray` or `DPTBinary` from `DPTBase.to_knx()` instead of `tuple[int, ...]`.
 
 # 2.7.0 IP Device Management 2023-03-15